backend/requirements.txt
Flask
flask-cors
Flask-SQLAlchemy
pymysql
flask-session
werkzeug
mysql-connector-python
